refactor: merged delete account
