+++ internal/handler/user_handler.go
@@ -10,6 +10,7 @@ import (
 	"github.com/PitiNarak/condormhub-backend/pkg/utils"
 	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
+	"github.com/google/uuid"
 )
 
 type UserHandler struct {
@@ -325,4 +326,30 @@ func (h *UserHandler) GetUserInfo(c *fiber.Ctx) error {
 	user := c.Locals("user").(*domain.User)
 	res := dto.Success(user)
 	return c.Status(fiber.StatusOK).JSON(res)
+}
+
+// DeleteAccount godoc
+// @Summary Delete a user account
+// @Description Delete a user account
+// @Tags user
+// @Security Bearer
+// @Accept json
+// @Produce json
+// @Success 204 "account successfully deleted"
+// @Failure 401 {object} dto.ErrorResponse "your request is unauthorized"
+// @Failure 500 {object} dto.ErrorResponse "cannot parse uuid or cannot delete user"
+// @Router /user/ [delete]
+func (h *UserHandler) DeleteAccount(c *fiber.Ctx) error {
+	userIDstr := c.Locals("userID").(string)
+	userID, err := uuid.Parse(userIDstr)
+	if err != nil {
+		return apperror.InternalServerError(err, "cannot parse uuid")
+	}
+
+	err = h.userService.DeleteAccount(userID)
+	if err != nil {
+		return err
+	}
+
+	return c.SendStatus(fiber.StatusNoContent)
 }
